# internal/controlapi/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Dorf Remote Control API",
    "version": "1.0.0",
    "description": "The authenticated, purpose-built remote boundary for one Dorf Deployment. A Session is the long-running resource; internal Tasks, AgentRuns, Actions, and providers are not exposed as resources. Timeline reads expose native Harness, Thread, and Turn references without granting control over them."
  },
  "servers": [
    {
      "url": "/",
      "description": "The connected Dorf Deployment"
    }
  ],
  "security": [
    {
      "clientBearer": []
    }
  ],
...
  "paths": {
...
    "/v1/auth/enrollments/redeem": {
      "post": {
        "operationId": "redeemEnrollment",
        "summary": "Redeem one short-lived Enrollment for a Client",
        "tags": ["Authentication"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RedeemRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Exact replay of an already redeemed Client binding",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Identity" }
              }
            }
          },
          "201": {
            "description": "Client binding created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Identity" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
    "/v1/me": {
      "get": {
        "operationId": "getIdentity",
        "summary": "Inspect the effective Principal and Client",
        "tags": ["Authentication"],
        "responses": {
          "200": {
            "description": "Authenticated identity",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Identity" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
    "/v1/profiles": {
      "get": {
        "operationId": "listProfiles",
        "summary": "List Sandbox profiles available for Session selection",
        "description": "Returns all configured profiles in name order, including unverified profiles. Verification reflects stored proof for the current definition, not a live availability check. Infrastructure configuration and credentials remain private.",
        "tags": ["Profiles"],
        "responses": {
          "200": {
            "description": "Narrow Sandbox profile summaries",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProfileList" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
    "/v1/sessions": {
      "get": {
        "operationId": "listSessions",
        "summary": "List one newest-first page of Sessions",
        "description": "This is keyset traversal of current Session facts, not a frozen snapshot. Pass next_cursor unchanged to retrieve the next page.",
        "tags": ["Sessions"],
        "parameters": [
          { "$ref": "#/components/parameters/PageLimit" },
          { "$ref": "#/components/parameters/PageCursor" }
        ],
        "responses": {
          "200": {
            "description": "One page of narrow Session summaries",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionList" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Problem" }
        }
      },
      "post": {
        "operationId": "createSession",
        "summary": "Admit or exactly replay a direct Session",
        "tags": ["Sessions"],
        "parameters": [
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateSessionRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Exact replay of the admitted Session",
            "headers": {
              "ETag": { "$ref": "#/components/headers/ETag" }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Session" }
              }
            }
          },
          "201": {
            "description": "Session admitted",
            "headers": {
              "ETag": { "$ref": "#/components/headers/ETag" }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Session" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
    "/v1/sessions/{session}": {
      "get": {
        "operationId": "getSession",
        "summary": "Read one canonical Session snapshot",
        "tags": ["Sessions"],
        "parameters": [
          { "$ref": "#/components/parameters/SessionID" },
          { "$ref": "#/components/parameters/IfNoneMatch" }
        ],
        "responses": {
          "200": {
            "description": "Current complete Session representation",
            "headers": {
              "ETag": { "$ref": "#/components/headers/ETag" }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Session" }
              }
            }
          },
          "304": {
            "description": "The supplied representation ETag is still current",
            "headers": {
              "ETag": { "$ref": "#/components/headers/ETag" }
            }
          },
          "default": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
    "/v1/sessions/{session}/messages/{message}/timeline": {
      "get": {
        "operationId": "readMessageTimeline",
        "summary": "Read completed replies and inputs in the Message's native turn",
        "description": "A passive read of the exact bound turn, including completed final replies before the turn settles. Entries preserve native order. Use thread_id, turn_id and index to identify retained entries; native_item_id is diagnostic only. Input message_id is present only when verified against Dorf custody. Missing binding or unavailable history returns timeline_unavailable.",
        "tags": [
          "Messages"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionID"
          },
          {
            "$ref": "#/components/parameters/MessageID"
          }
        ],
        "responses": {
          "200": {
            "description": "Completed conversation entries available at read time",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageTimeline"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/v1/sessions/{session}/timeline": {
      "get": {
        "operationId": "readSessionTimeline",
        "summary": "Read one native conversation turn without starting work",
        "description": "Returns original userMessage and agentMessage objects in native order. Omitted turn_id selects the latest started turn once. Unavailable, unsupported, oversized, or incomplete history fails without truncation.",
        "tags": [
          "Sessions"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/SessionID"
          },
          {
            "name": "turn_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Native conversation turn available at read time",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Timeline"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/v1/sessions/{session}/watch": {
      "get": {
        "operationId": "watchSession",
        "summary": "Watch newer complete Session snapshots",
        "description": "A non-durable server-sent event stream. Each snapshot event carries an id equal to the canonical representation digest and data containing one complete Session JSON object. Keepalives are comments. Reconnect by reading the canonical Session or by supplying the last event id; the stream closes for bounded reauthentication.",
        "tags": ["Sessions"],
        "parameters": [
          { "$ref": "#/components/parameters/SessionID" },
          { "$ref": "#/components/parameters/SSEAccept" },
          { "$ref": "#/components/parameters/LastEventID" }
        ],
        "responses": {
          "200": {
            "description": "Server-sent snapshot events and keepalive comments",
            "content": {
              "text/event-stream": {
                "schema": { "type": "string" },
                "x-dorf-event": {
                  "name": "snapshot",
                  "id": "SHA-256 hex digest of data",
                  "dataSchema": { "$ref": "#/components/schemas/Session" }
                }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
...
    "/v1/sessions/{session}/messages/{message}": {
      "get": {
        "operationId": "getSessionMessage",
        "summary": "Inspect one accepted Message delivery",
        "tags": ["Messages"],
        "parameters": [
          { "$ref": "#/components/parameters/SessionID" },
          { "$ref": "#/components/parameters/MessageID" }
        ],
        "responses": {
          "200": {
            "description": "Current retained Message delivery",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Message" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/Problem" }
        },
        "description": " Use latest as the Message selector to inspect the latest settled reply in the main Sandbox."
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "clientBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "One opaque, individually expiring and revocable Dorf Client credential."
      }
    },
    "parameters": {
...
      "PageLimit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Maximum summaries in this page",
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 }
      },
...
    },
...
    "schemas": {
...
      "RedeemRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["enrollment_code", "client_name", "credential"],
        "properties": {
          "enrollment_code": { "type": "string", "minLength": 1 },
          "client_name": { "type": "string", "minLength": 1 },
          "credential": {
            "type": "string",
            "minLength": 1,
            "writeOnly": true,
            "description": "Client-generated secret; Dorf stores only its digest and never returns it."
          }
        }
      },
...
      "Identity": {
        "type": "object",
        "additionalProperties": false,
        "required": ["principal", "client"],
        "properties": {
          "principal": { "$ref": "#/components/schemas/Principal" },
          "client": { "$ref": "#/components/schemas/Client" }
        }
      },
...
    }
  }
}